Moved snmp and ipc clients to platform, updates to ignite code after rebase.
Small update of platform feature set.
