+++ minion/snmp/pom.xml
@@ -16,7 +16,6 @@
   <modules>
     <module>snmp4j</module>
     <module>snmp-ipc</module>
-    <module>snmp-ipc-client</module>
   </modules>
 
 </project>
