+++ platform/ignite/src/main/java/org/opennms/horizon/core/ignite/rpc/IgniteRpcClient.java
@@ -7,7 +7,6 @@ import org.apache.ignite.client.IgniteClient;
 import org.apache.ignite.client.IgniteClientFuture;
 import org.opennms.cloud.grpc.minion.RpcRequestProto;
 import org.opennms.cloud.grpc.minion.RpcResponseProto;
-import org.opennms.horizon.shared.ignite.remoteasync.MinionRouterService;
 import org.opennms.horizon.shared.ipc.rpc.api.RequestBuilder;
 import org.opennms.horizon.shared.ipc.rpc.api.RpcClient;
 import org.opennms.horizon.shared.ipc.rpc.api.RpcClientFactory.Deserializer;
@@ -24,7 +23,7 @@ public class IgniteRpcClient<T extends Message> implements RpcClient<T> {
     @Override
     public CompletableFuture<T> execute(RpcRequestProto request) {
         IgniteClientFuture<RpcResponseProto> future = client.compute()
-            .executeAsync2(MinionRouterService.IGNITE_SERVICE_NAME, request);
+            .executeAsync2("echoRoutingTask", request);
         return future.toCompletableFuture().thenApply(response -> {
             try {
                 return deserializer.deserialize(response);
