+++ platform/ipc/icmp-ipc-client/pom.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+  <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>org.opennms.horizon.ipc</groupId>
+        <artifactId>ipc</artifactId>
+        <version>0.1.0-SNAPSHOT</version>
+    </parent>
+
+  <artifactId>icmp-ipc-client</artifactId>
+  <name>OpenNMS Horizon :: IPC :: ICMP IPC Client</name>
+  <packaging>bundle</packaging>
+
+  <dependencies>
+    <dependency>
+      <groupId>org.opennms.horizon.shared</groupId>
+      <artifactId>icmp-api</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.opennms.horizon.shared</groupId>
+      <artifactId>ipc-api</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.opennms.horizon.shared</groupId>
+      <artifactId>ipc-grpc-contract</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+  </dependencies>
+
+</project>
