+++ platform/ipc/icmp-ipc-client/src/main/resources/OSGI-INF/blueprint/icmp-ipc-client.xml
@@ -1,0 +1,16 @@
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="
+        http://www.osgi.org/xmlns/blueprint/v1.0.0
+        https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+    ">
+
+    <reference id="rpcClientFactory" interface="org.opennms.horizon.shared.ipc.rpc.api.RpcClientFactory"/>
+
+    <service interface="org.opennms.netmgt.icmp.proxy.LocationAwarePingClient">
+        <bean class="org.opennms.horizon.icmp.ipc.client.internal.LocationAwarePingClientImpl">
+            <argument ref="rpcClientFactory" />
+        </bean>
+    </service>
+
+</blueprint>
