+++ platform/ipc/pom.xml
@@ -21,6 +21,8 @@
     <module>shell</module>
     <module>echo</module>
     <module>twin</module>
+    <module>icmp-ipc-client</module>
+    <module>snmp-ipc-client</module>
   </modules>
 
 </project>
