+++ platform/ipc/snmp-ipc-client/src/main/java/org/opennms/horizon/snmp/ipc/client/internal/SNMPWalkBuilder.java
@@ -1,0 +1,61 @@
+/*******************************************************************************
+ * This file is part of OpenNMS(R).
+ *
+ * Copyright (C) 2016 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2016 The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * OpenNMS(R) is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with OpenNMS(R).  If not, see:
+ *      http://www.gnu.org/licenses/
+ *
+ * For more information contact:
+ *     OpenNMS(R) Licensing <license@opennms.org>
+ *     http://www.opennms.org/
+ *     http://www.opennms.com/
+ *******************************************************************************/
+
+package org.opennms.horizon.snmp.ipc.client.internal;
+
+import java.util.Collections;
+import java.util.List;
+import org.opennms.horizon.grpc.snmp.contract.SnmpMultiResponse;
+import org.opennms.horizon.grpc.snmp.contract.SnmpWalkRequest;
+import org.opennms.horizon.grpc.snmp.contract.SnmpWalkRequest.Builder;
+import org.opennms.horizon.shared.snmp.SnmpAgentConfig;
+import org.opennms.horizon.shared.snmp.SnmpObjId;
+import org.opennms.horizon.shared.snmp.SnmpResult;
+import org.opennms.horizon.shared.snmp.SnmpValueFactory;
+
+public class SNMPWalkBuilder extends AbstractSNMPRequestBuilder<List<SnmpResult>> {
+
+    public SNMPWalkBuilder(LocationAwareSnmpClientRpcImpl client, SnmpAgentConfig agent, SnmpValueFactory valueFactory, List<SnmpObjId> oids) {
+        super(client, agent, valueFactory, Collections.emptyList(), buildWalkRequests(oids));
+    }
+
+    @Override
+    protected List<SnmpResult> processResponse(SnmpMultiResponse response) {
+        return response.getResponsesList().stream()
+            .findFirst()
+            .map(this::mapResults)
+            .orElse(Collections.emptyList());
+    }
+
+    private static List<SnmpWalkRequest> buildWalkRequests(List<SnmpObjId> oids) {
+        Builder walkRequest = SnmpWalkRequest.newBuilder();
+        oids.forEach(oid -> walkRequest.addOids(oid.toString()));
+        return Collections.singletonList(walkRequest.build());
+    }
+
+}
