+++ platform/ipc/snmp-ipc-client/src/main/java/org/opennms/horizon/snmp/ipc/client/internal/SNMPWalkWithTrackerBuilder.java
@@ -1,0 +1,71 @@
+/*******************************************************************************
+ * This file is part of OpenNMS(R).
+ *
+ * Copyright (C) 2016 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2016 The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * OpenNMS(R) is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with OpenNMS(R).  If not, see:
+ *      http://www.gnu.org/licenses/
+ *
+ * For more information contact:
+ *     OpenNMS(R) Licensing <license@opennms.org>
+ *     http://www.opennms.org/
+ *     http://www.opennms.com/
+ *******************************************************************************/
+
+package org.opennms.horizon.snmp.ipc.client.internal;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
+import org.opennms.horizon.grpc.snmp.contract.SnmpMultiResponse;
+import org.opennms.horizon.grpc.snmp.contract.SnmpWalkRequest;
+import org.opennms.horizon.shared.snmp.CollectionTracker;
+import org.opennms.horizon.shared.snmp.SnmpAgentConfig;
+import org.opennms.horizon.shared.snmp.SnmpValueFactory;
+import org.opennms.horizon.shared.snmp.proxy.WalkResponse;
+
+public class SNMPWalkWithTrackerBuilder extends AbstractSNMPRequestBuilder<CollectionTracker> {
+
+    private final CollectionTracker m_tracker;
+
+    public SNMPWalkWithTrackerBuilder(LocationAwareSnmpClientRpcImpl client, SnmpAgentConfig agent, SnmpValueFactory valueFactory, CollectionTracker tracker) {
+        super(client, agent,  valueFactory, Collections.emptyList(), buildWalkRequests(tracker));
+        m_tracker = tracker;
+    }
+
+    private static List<SnmpWalkRequest> buildWalkRequests(CollectionTracker tracker) {
+        return tracker.getWalkRequests().stream()
+            .map(req -> SnmpWalkRequest.newBuilder()
+                .setCorrelationId(req.getCorrelationId())
+                .addAllOids(Collections.singletonList(req.getBaseOid().toString()))
+                .setMaxRepetitions(req.getMaxRepetitions())
+                .setInstance(req.getInstance().toString())
+                .build()
+            ).collect(Collectors.toList());
+    }
+
+
+    @Override
+    protected CollectionTracker processResponse(SnmpMultiResponse response) {
+        final List<WalkResponse> responses = response.getResponsesList().stream()
+            .map(res -> new WalkResponse(mapResults(res), res.getCorrelationId()))
+            .collect(Collectors.toList());
+        m_tracker.handleWalkResponses(responses);
+        return m_tracker;
+    }
+
+}
