+++ platform/ipc/snmp-ipc-client/src/main/resources/OSGI-INF/blueprint/snmp-ipc-client.xml
@@ -1,0 +1,18 @@
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="
+        http://www.osgi.org/xmlns/blueprint/v1.0.0
+        https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+    ">
+
+    <reference id="rpcClientFactory" interface="org.opennms.horizon.shared.ipc.rpc.api.RpcClientFactory"/>
+    <reference id="snmpValueFactory" interface="org.opennms.horizon.shared.snmp.SnmpValueFactory"/>
+
+    <service interface="org.opennms.horizon.shared.snmp.proxy.LocationAwareSnmpClient">
+        <bean class="org.opennms.horizon.snmp.ipc.client.internal.LocationAwareSnmpClientRpcImpl">
+            <argument ref="rpcClientFactory" />
+            <argument ref="snmpValueFactory" />
+        </bean>
+    </service>
+
+</blueprint>
