+++ platform/provision/ignite-service/pom.xml
@@ -37,11 +37,6 @@
             <artifactId>ipc-grpc-contract</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>org.opennms.horizon.minion.gateway</groupId>
-            <artifactId>minion-gateway-ignite-detector-client</artifactId>
-            <version>${project.version}</version>
-        </dependency>
 
         <dependency>
             <groupId>org.opennms.horizon.shared</groupId>
